backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
channels==4.0.0
django-cors-headers==4.3.0
faker==20.1.0
psycopg2-binary==2.9.9
daphne==4.0.0
python-decouple==3.8
requests==2.31.0
websockets==12.0
